feat!: [Java] fix up #802: `SupportedDevices`を移動
